Add authorization, audit, logging integration.

+++ EHospital.Medications.WebAPI/Controllers/DrugsController.cs
@@ -5,21 +5,44 @@ using Microsoft.AspNetCore.Mvc;
 using EHospital.Medications.BusinessLogic.Contracts;
 using EHospital.Medications.BusinessLogic.Services;
 using EHospital.Medications.Model;
+using EHospital.Shared.AuditTrail;
+using EHospital.Shared.AuditTrail.Models;
+using EHospital.Shared.Authorization;
+using EHospital.Shared.Configuration;
+using EHospital.Shared.Logging;
+using EHospital.Shared.Logging.Models;
+using Microsoft.AspNetCore.Http.Extensions;
 
 namespace EHospital.Medications.WebAPI.Controllers
 {
-    // TODO: Logging
-    // TODO: Mapper
-    // TODO: UnitTest Coverage
-    // TODO: UpdatePostman
-
     /// <summary>
     /// Represents drug controller.
     /// </summary>
+    /// <seealso cref="Microsoft.AspNetCore.Mvc.ControllerBase" />
     /// <seealso cref="ControllerBase" />
     [ApiController]
     public class DrugsController : ControllerBase
     {
+        /// <summary>
+        /// The authentication details provider.
+        /// </summary>
+        private readonly IAuthDetailsProvider authDetailsProvider;
+
+        /// <summary>
+        /// The configuration provider.
+        /// </summary>
+        private readonly IConfigurationProvider configurationProvider;
+
+        /// <summary>
+        /// The logging provider.
+        /// </summary>
+        private readonly ILoggingProvider loggingProvider;
+
+        /// <summary>
+        /// The audit trail provider
+        /// </summary>
+        private readonly IAuditTrailProvider auditTrailProvider;
+
         /// <summary>
         /// Represents default route, which contain all REST requests.
         /// </summary>
@@ -34,9 +57,22 @@ namespace EHospital.Medications.WebAPI.Controllers
         /// Initializes a new instance of the <see cref="DrugsController"/> class.
         /// </summary>
         /// <param name="service">The drug service.</param>
-        public DrugsController(IDrugService service)
+        /// <param name="authDetailsProvider">The authentication details provider.</param>
+        /// <param name="configurationProvider">The configuration provider.</param>
+        /// <param name="loggingProvider">The logging provider.</param>
+        /// <param name="auditTrailProvider">The audit trail provider.</param>
+        public DrugsController(
+            IDrugService service,
+            IAuthDetailsProvider authDetailsProvider,
+            IConfigurationProvider configurationProvider,
+            ILoggingProvider loggingProvider,
+            IAuditTrailProvider auditTrailProvider)
         {
             this.service = service;
+            this.authDetailsProvider = authDetailsProvider;
+            this.configurationProvider = configurationProvider;
+            this.loggingProvider = loggingProvider;
+            this.auditTrailProvider = auditTrailProvider;
         }
 
         /// <summary>
@@ -54,12 +90,30 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 IEnumerable<Drug> result = await this.service.GetAllAsync();
                 return this.Ok(result);
             }
             catch (NoContentException)
             {
                 return this.NoContent();
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -82,12 +136,30 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 IEnumerable<Drug> result = await this.service.GetAllByNameAsync(name);
                 return this.Ok(result);
             }
             catch (NoContentException)
             {
                 return this.NoContent();
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -108,12 +180,30 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 Drug result = await this.service.GetByIdAsync(id);
                 return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.NotFound(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -134,19 +224,49 @@ namespace EHospital.Medications.WebAPI.Controllers
         [HttpPost(DEFAULT_ROUTE + "add")]
         public async Task<IActionResult> AddDrug([FromBody] Drug drug)
         {
-            if (!this.ModelState.IsValid)
-            {
-                return this.ValidationProblem(this.ModelState);
-            }
-
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
+                if (!this.ModelState.IsValid)
+                {
+                    return this.ValidationProblem(this.ModelState);
+                }
+
                 Drug result = await this.service.AddAsync(drug);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl,this.Request.Headers["Authorization"], new AuditTrailModel
+                    {
+                        ActionItem = result.GetType().Name,
+                        ActionType = ActionMode.Create,
+                        Description = $"Drug has been created with id = {result.Id}.",
+                        UserId = authInfo.Id,
+                        ItemId = result.Id,
+                        ItemState = result,
+                        Module = "Medications.API"
+                    });
+
                 return this.Created(DEFAULT_ROUTE, result.Id);
             }
             catch (ArgumentException ex)
             {
                 return this.BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -169,19 +289,49 @@ namespace EHospital.Medications.WebAPI.Controllers
         [HttpPut(DEFAULT_ROUTE + "edit/{id}")]
         public async Task<IActionResult> EditDrug(int id, [FromBody] Drug drug)
         {
-            if (!this.ModelState.IsValid)
-            {
-                return this.ValidationProblem(this.ModelState);
-            }
-
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
+                if (!this.ModelState.IsValid)
+                {
+                    return this.ValidationProblem(this.ModelState);
+                }
+
                 Drug result = await this.service.UpdateAsync(id, drug);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"], new AuditTrailModel
+                {
+                    ActionItem = result.GetType().Name,
+                    ActionType = ActionMode.Update,
+                    Description = $"Drug with id = {result.Id} has been updated.",
+                    UserId = authInfo.Id,
+                    ItemId = result.Id,
+                    ItemState = result,
+                    Module = "Medications.API"
+                });
+
                 return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -201,12 +351,42 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 Drug result = await this.service.DeleteAsync(id);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"], new AuditTrailModel
+                {
+                    ActionItem = result.GetType().Name,
+                    ActionType = ActionMode.Delete,
+                    Description = $"Drug with id = {result.Id} has been soft deleted.",
+                    UserId = authInfo.Id,
+                    ItemId = result.Id,
+                    ItemState = result,
+                    Module = "Medications.API"
+                });
+
                 return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
     }
