+++ EHospital.Medications.WebAPI/Controllers/PrescriptionsController.cs
@@ -4,6 +4,13 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using EHospital.Medications.BusinessLogic.Contracts;
 using EHospital.Medications.Model;
+using EHospital.Shared.AuditTrail;
+using EHospital.Shared.AuditTrail.Models;
+using EHospital.Shared.Authorization;
+using EHospital.Shared.Configuration;
+using EHospital.Shared.Logging;
+using EHospital.Shared.Logging.Models;
+using Microsoft.AspNetCore.Http.Extensions;
 
 namespace EHospital.Medications.WebAPI.Controllers
 {
@@ -14,6 +21,26 @@ namespace EHospital.Medications.WebAPI.Controllers
     [ApiController]
     public class PrescriptionsController : ControllerBase
     {
+        /// <summary>
+        /// The authentication details provider.
+        /// </summary>
+        private readonly IAuthDetailsProvider authDetailsProvider;
+
+        /// <summary>
+        /// The configuration provider.
+        /// </summary>
+        private readonly IConfigurationProvider configurationProvider;
+
+        /// <summary>
+        /// The logging provider.
+        /// </summary>
+        private readonly ILoggingProvider loggingProvider;
+
+        /// <summary>
+        /// The audit trail provider
+        /// </summary>
+        private readonly IAuditTrailProvider auditTrailProvider;
+
         /// <summary>
         /// Represents default route, which contain all REST requests.
         /// </summary>
@@ -28,9 +55,22 @@ namespace EHospital.Medications.WebAPI.Controllers
         /// Initializes a new instance of the <see cref="PrescriptionsController"/> class.
         /// </summary>
         /// <param name="service">The prescription service.</param>
-        public PrescriptionsController(IPrescriptionService service)
+        /// <param name="authDetailsProvider">The authentication details provider.</param>
+        /// <param name="configurationProvider">The configuration provider.</param>
+        /// <param name="loggingProvider">The logging provider.</param>
+        /// <param name="auditTrailProvider">The audit trail provider.</param>
+        public PrescriptionsController(
+            IPrescriptionService service,
+            IAuthDetailsProvider authDetailsProvider,
+            IConfigurationProvider configurationProvider,
+            ILoggingProvider loggingProvider,
+            IAuditTrailProvider auditTrailProvider)
         {
             this.service = service;
+            this.authDetailsProvider = authDetailsProvider;
+            this.configurationProvider = configurationProvider;
+            this.loggingProvider = loggingProvider;
+            this.auditTrailProvider = auditTrailProvider;
         }
 
         /// <summary>
@@ -53,12 +93,30 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 IEnumerable<PrescriptionDetails> result = await this.service.GetPrescriptionsDetails(patientId);
                 return this.Ok(result);
             }
             catch (ArgumentException)
             {
                 return this.NoContent();
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -81,12 +139,30 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 Prescription result = await this.service.GetByIdAsync(id);
                 return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.NotFound(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -109,12 +185,30 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
                 PrescriptionGuide result = await this.service.GetGuideById(id);
                 return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.NotFound(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -135,13 +229,47 @@ namespace EHospital.Medications.WebAPI.Controllers
         [HttpPost(DEFAULT_ROUTE + "add")]
         public async Task<IActionResult> AddPrescription([FromBody] Prescription prescription)
         {
-            if (!this.ModelState.IsValid)
+            try
             {
-                return this.ValidationProblem(this.ModelState);
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
+                if (!this.ModelState.IsValid)
+                {
+                    return this.ValidationProblem(this.ModelState);
+                }
+
+                Prescription result = await this.service.AddAsync(prescription);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"], new AuditTrailModel
+                {
+                    ActionItem = result.GetType().Name,
+                    ActionType = ActionMode.Create,
+                    Description = $"Prescription has been created with id = {result.Id}.",
+                    UserId = authInfo.Id,
+                    ItemId = result.Id,
+                    ItemState = result,
+                    Module = "Medications.API"
+                });
+
+                return this.Created(DEFAULT_ROUTE, prescription.Id);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
 
-            Prescription result = await this.service.AddAsync(prescription);
-            return this.Created(DEFAULT_ROUTE, prescription.Id);
         }
 
         /// <summary>
@@ -163,19 +291,49 @@ namespace EHospital.Medications.WebAPI.Controllers
         [HttpPut(DEFAULT_ROUTE + "edit/{id}")]
         public async Task<IActionResult> EditPrescription(int id, [FromBody] Prescription prescription)
         {
-            if (!this.ModelState.IsValid)
-            {
-                return this.ValidationProblem(this.ModelState);
-            }
-
             try
             {
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
+                if (!this.ModelState.IsValid)
+                {
+                    return this.ValidationProblem(this.ModelState);
+                }
+
                 Prescription result = await this.service.UpdateAsync(id, prescription);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"], new AuditTrailModel
+                {
+                    ActionItem = result.GetType().Name,
+                    ActionType = ActionMode.Update,
+                    Description = $"Prescription with id = {result.Id} has been updated.",
+                    UserId = authInfo.Id,
+                    ItemId = result.Id,
+                    ItemState = result,
+                    Module = "Medications.API"
+                });
+
                 return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -195,12 +353,42 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
-                Prescription prescription = await this.service.DeleteAsync(id);
-                return this.Ok(prescription);
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
+                Prescription result = await this.service.DeleteAsync(id);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"], new AuditTrailModel
+                {
+                    ActionItem = result.GetType().Name,
+                    ActionType = ActionMode.Delete,
+                    Description = $"Prescription with id = {result.Id} has been soft deleted.",
+                    UserId = authInfo.Id,
+                    ItemId = result.Id,
+                    ItemState = result,
+                    Module = "Medications.API"
+                });
+
+                return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
 
@@ -221,12 +409,42 @@ namespace EHospital.Medications.WebAPI.Controllers
         {
             try
             {
-                Prescription prescription = await this.service.UpdateStatusAsync(id);
-                return this.Ok(prescription);
+                var authInfo = await this.authDetailsProvider.GetUserAuthInfoAsync(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"]);
+                if (authInfo == null)
+                {
+                    return this.Unauthorized();
+                }
+
+                Prescription result = await this.service.UpdateStatusAsync(id);
+
+                await this.auditTrailProvider.LogAuditTrailMessage(this.configurationProvider.BaseUrl, this.Request.Headers["Authorization"], new AuditTrailModel
+                {
+                    ActionItem = result.GetType().Name,
+                    ActionType = ActionMode.Update,
+                    Description = $"Prescription status has been updated. Id = {result.Id}.",
+                    UserId = authInfo.Id,
+                    ItemId = result.Id,
+                    ItemState = result,
+                    Module = "Medications.API"
+                });
+
+                return this.Ok(result);
             }
             catch (ArgumentException ex)
             {
                 return this.BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                await this.loggingProvider.LogErrorMessage(this.configurationProvider.BaseUrl, new LoggingMessage
+                {
+                    ErrorMessage = ex.Message,
+                    Exception = ex,
+                    RequestType = HttpContext.Request.Method,
+                    RequestUri = HttpContext.Request.GetDisplayUrl()
+                });
+
+                return this.BadRequest("Some error was thrown:" + ex.Message);
             }
         }
     }
